syntax fixes
